refactor(admin): extract shared CapabilityPicker used by Role + MCP dialogs
Move the role dialog's capability checklist into a shared, presentational
`@admin/shared/CapabilityPicker` (+ co-located CAPABILITY_META). RoleDialog and
the MCP connector dialog both consume it, so they stay visually consistent and
the markup/CSS lives in one place. UsersPage.module.css keeps only the role
identity grid; users/utils/capabilities.ts keeps only the role grouping.
