+++ src/admin/shared/CapabilityPicker/CapabilityPicker.module.css
@@ -1,0 +1,191 @@
+/* Shared capability picker.
+
+   Layout: a single flow inside a Dialog body. No inner scroll — the Dialog
+   already provides the scroll container, so a second scrollbar here would be a
+   UX bug. The sticky header keeps the running selected count and master toggle
+   in view while scrolling. Each group is a card (`--bg-surface-2` on the Dialog
+   body surface) following the project's card-surface pattern.
+
+   The header's negative margin + matching padding bleed it to the Dialog body
+   edges; it assumes the Dialog body padding is `--space-2xl` (the shared Dialog
+   primitive). */
+.capabilityPicker {
+    display: grid;
+    gap: var(--space-l);
+    margin-top: var(--space-3xs);
+}
+
+.capabilityPickerHeader {
+    position: sticky;
+    top: -16px;
+    z-index: 1;
+    display: grid;
+    grid-template-columns: minmax(0, 1fr) auto;
+    align-items: center;
+    gap: var(--space-l);
+    margin: calc(var(--space-2xl) * -1) calc(var(--space-2xl) * -1) 0;
+    padding: var(--space-l) var(--space-2xl);
+    background: var(--bg-surface);
+    border-bottom: 1px solid var(--border);
+}
+
+.capabilityPickerSummary {
+    display: flex;
+    flex-wrap: wrap;
+    align-items: baseline;
+    gap: var(--space-3xs) var(--space-xl);
+    min-width: 0;
+}
+
+.capabilityPickerTitle {
+    margin: 0;
+    color: var(--text-bright);
+    font-size: var(--text-m);
+    font-weight: 700;
+    line-height: 1.2;
+}
+
+.capabilityPickerCount {
+    margin: 0;
+    color: var(--text-subtle);
+    font-size: var(--text-xs);
+    font-variant-numeric: tabular-nums;
+    line-height: 1.2;
+}
+
+.capabilityPickerCount strong {
+    color: var(--text-bright);
+    font-weight: 700;
+}
+
+.capabilityGroups {
+    display: grid;
+    gap: var(--space-m);
+}
+
+.capabilityGroup {
+    display: grid;
+    gap: var(--space-m);
+    padding: var(--space-l);
+    border-radius: var(--panel-radius);
+    background: var(--bg-surface-2);
+}
+
+.capabilityGroupHeader {
+    display: grid;
+    grid-template-columns: minmax(0, 1fr) auto;
+    align-items: center;
+    gap: var(--space-l);
+}
+
+.capabilityGroupHeading {
+    display: flex;
+    align-items: center;
+    gap: var(--space-s);
+    min-width: 0;
+}
+
+.capabilityGroupHeader h4 {
+    margin: 0;
+    color: var(--text-bright);
+    font-size: var(--text-s);
+    font-weight: 700;
+    line-height: 1.2;
+    letter-spacing: 0.01em;
+}
+
+.capabilityGroupCount {
+    display: inline-flex;
+    align-items: center;
+    height: 18px;
+    padding: 0 var(--space-xs);
+    border-radius: var(--radius);
+    background: var(--bg-surface-3);
+    color: var(--text-subtle);
+    font-size: var(--text-2xs);
+    font-weight: 700;
+    font-variant-numeric: tabular-nums;
+    line-height: 1;
+}
+
+.capabilityGroupCount[data-state="full"] {
+    background: var(--success-10);
+    color: var(--text-bright);
+}
+
+.capabilityGroupCount[data-state="partial"] {
+    color: var(--text);
+}
+
+.capabilityList {
+    display: grid;
+    grid-template-columns: repeat(2, minmax(0, 1fr));
+    gap: var(--space-3xs) var(--space-s);
+    margin: 0;
+    padding: 0;
+    list-style: none;
+}
+
+.capabilityItem {
+    display: block;
+}
+
+.capabilityRow {
+    display: grid;
+    grid-template-columns: auto minmax(0, 1fr);
+    align-items: flex-start;
+    gap: var(--space-m);
+    padding: var(--space-s) var(--space-m);
+    border-radius: var(--radius);
+    cursor: pointer;
+    transition: background-color 120ms ease;
+}
+
+.capabilityRow:hover {
+    background: var(--bg-surface-3);
+}
+
+.capabilityItem[data-checked="true"] .capabilityRow {
+    background: color-mix(in srgb, var(--bg-surface-3) 60%, transparent);
+}
+
+.capabilityRow:has(input:focus-visible) {
+    outline: 2px solid var(--canvas-selection-ring-color);
+    outline-offset: 0;
+}
+
+.capabilityRow:has(input:disabled) {
+    cursor: default;
+}
+
+.capabilityRow:has(input:disabled):hover {
+    background: transparent;
+}
+
+.capabilityRowText {
+    display: grid;
+    gap: var(--space-4xs);
+    min-width: 0;
+}
+
+.capabilityRowLabel {
+    color: var(--text-bright);
+    font-size: var(--text-s);
+    font-weight: 600;
+    line-height: 1.3;
+    overflow-wrap: anywhere;
+}
+
+.capabilityRowDescription {
+    color: var(--text-subtle);
+    font-size: var(--text-xs);
+    font-weight: 500;
+    line-height: 1.35;
+    overflow-wrap: anywhere;
+}
+
+@media (max-width: 620px) {
+    .capabilityList {
+        grid-template-columns: 1fr;
+    }
+}
