+++ src/admin/shared/CapabilityPicker/index.ts
@@ -1,0 +1,2 @@
+export { CapabilityPicker, type CapabilityPickerGroup } from './CapabilityPicker'
+export { CAPABILITY_META, capabilityLabel, type CapabilityMeta } from './capabilityMeta'
